requirements.txt
itsdangerous
pandas==1.5.2
google-api-python-client
google-api-core
google-auth
google-auth-httplib2
google-cloud-storage
google-cloud-bigquery==2.34.3
pycaret==2.3.10
pyarrow
scikit-learn==0.23.2